Added getOrdersWithApprovalWarnings to limit orders
