+++ src/components/@widgets/MyLimitOrdersWidget/subcomponents/MyLimitOrdersList/MyLimitOrdersList.tsx
@@ -5,17 +5,22 @@ import { TokenInfo } from "@airswap/utils";
 import * as ethers from "ethers";
 
 import { useAppSelector } from "../../../../../app/hooks";
+import { isTokenInfo } from "../../../../../entities/AppTokenInfo/AppTokenInfoHelpers";
 import { DelegateRule } from "../../../../../entities/DelegateRule/DelegateRule";
+import { Allowances } from "../../../../../features/balances/balancesTypes";
 import { selectAllTokenInfo } from "../../../../../features/metadata/metadataSlice";
 import { OrdersSortType } from "../../../../../types/ordersSortType";
 import { MyOrder } from "../../../MyOrdersWidget/entities/MyOrder";
+import { getOrdersWithApprovalWarnings } from "../../../MyOrdersWidget/helpers";
 import { StyledMyLimitOrdersList } from "./MyLimitOrdersList.styles";
 import { getDelegateRuleDataAndTransformToMyOrder } from "./helpers";
 
 interface MyLimitOrdersListProps {
+  isAllowancesLoading: boolean;
   activeCancellationId?: string;
   activeSortType: OrdersSortType;
   activeTokens: TokenInfo[];
+  allowances: Allowances;
   delegateRules: DelegateRule[];
   sortTypeDirection: Record<OrdersSortType, boolean>;
   library: ethers.providers.BaseProvider;
@@ -25,8 +30,10 @@ interface MyLimitOrdersListProps {
 }
 
 const MyLimitOrdersList: FC<MyLimitOrdersListProps> = ({
+  isAllowancesLoading,
   activeCancellationId,
   activeSortType,
+  allowances,
   delegateRules,
   library,
   sortTypeDirection,
@@ -46,7 +53,14 @@ const MyLimitOrdersList: FC<MyLimitOrdersListProps> = ({
       )
     );
 
-    setOrders(newOrders);
+    const ordersWithApprovalWarnings = getOrdersWithApprovalWarnings(
+      newOrders.filter(
+        (order) => order.signerToken && isTokenInfo(order.signerToken)
+      ),
+      allowances.delegate.values
+    );
+
+    setOrders(ordersWithApprovalWarnings);
     setIsLoading(false);
   }, [delegateRules, activeTokens, activeCancellationId]);
 
@@ -72,13 +86,18 @@ const MyLimitOrdersList: FC<MyLimitOrdersListProps> = ({
   }, [activeCancellationId]);
 
   useEffect(() => {
-    callGetOrders();
+    if (
+      allowances.swapERC20.status === "idle" &&
+      allowances.swap.status === "idle"
+    ) {
+      callGetOrders();
+    }
   }, [delegateRules]);
 
   return (
     <StyledMyLimitOrdersList
       hasFilledColumn
-      isLoading={isLoading}
+      isLoading={isLoading || isAllowancesLoading}
       activeSortType={activeSortType}
       orders={orders}
       sortTypeDirection={sortTypeDirection}
