+++ src/components/@widgets/MyOrdersWidget/helpers/index.ts
@@ -1,7 +1,11 @@
 import { BigNumber } from "bignumber.js";
+import { BigNumber as EthersBigNumber } from "ethers";
 import i18n from "i18next";
 
+import { getTokenId } from "../../../../entities/AppTokenInfo/AppTokenInfoHelpers";
+import { BalanceValues } from "../../../../features/balances/balancesSlice";
 import { OrderStatus } from "../../../../types/orderStatus";
+import { MyOrder } from "../entities/MyOrder";
 
 export const getTokenAmountWithDecimals = (
   amount: string,
@@ -28,4 +32,47 @@ export const getOrderStatusTranslation = (status: OrderStatus): string => {
   }
 
   return i18n.t("common.active");
+};
+
+const getOrdersTotalApprovalAmount = (orders: MyOrder[]) => {
+  return orders.reduce((acc, order) => {
+    if (!order.signerToken || order.status !== OrderStatus.open) {
+      return acc;
+    }
+
+    const tokenId = getTokenId(order.signerToken);
+
+    const currentAmount = acc[tokenId] || "0";
+    acc[tokenId] = EthersBigNumber.from(currentAmount)
+      .add(EthersBigNumber.from(order.signerAmount))
+      .toString();
+
+    return acc;
+  }, {} as BalanceValues);
+};
+
+export const getOrdersWithApprovalWarnings = (
+  orders: MyOrder[],
+  allowances: BalanceValues
+) => {
+  const tokenApprovals = getOrdersTotalApprovalAmount(orders);
+
+  return orders.map((order) => {
+    if (!order.signerToken) {
+      return order;
+    }
+
+    const tokenId = getTokenId(order.signerToken);
+    const approvedAmount = allowances[tokenId] || "0";
+    const tokensAmount = tokenApprovals[tokenId] || "0";
+
+    const hasAllowanceWarning = EthersBigNumber.from(approvedAmount).lt(
+      EthersBigNumber.from(tokensAmount)
+    );
+
+    return {
+      ...order,
+      hasAllowanceWarning,
+    } as MyOrder;
+  });
 };
