+++ src/components/@widgets/MyOtcOrdersWidget/subcomponents/MyOtcOrdersList/helpers/index.ts
@@ -104,46 +104,3 @@ export const getFullOrderDataAndTransformToOrder = async (
 
   return transformFullOrderToMyOrder(order, status, signerToken, senderToken);
 };
-
-export const getOrdersTotalApprovalAmount = (orders: MyOrder[]) => {
-  return orders.reduce((acc, order) => {
-    if (!order.signerToken || order.status !== OrderStatus.open) {
-      return acc;
-    }
-
-    const tokenId = getTokenId(order.signerToken);
-
-    const currentAmount = acc[tokenId] || "0";
-    acc[tokenId] = BigNumber.from(currentAmount)
-      .add(BigNumber.from(order.signerAmount))
-      .toString();
-
-    return acc;
-  }, {} as BalanceValues);
-};
-
-export const getOrdersWithApprovalWarnings = (
-  orders: MyOrder[],
-  allowances: BalanceValues
-) => {
-  const tokenApprovals = getOrdersTotalApprovalAmount(orders);
-
-  return orders.map((order) => {
-    if (!order.signerToken) {
-      return order;
-    }
-
-    const tokenId = getTokenId(order.signerToken);
-    const approvedAmount = allowances[tokenId] || "0";
-    const tokensAmount = tokenApprovals[tokenId] || "0";
-
-    const hasAllowanceWarning = BigNumber.from(approvedAmount).lt(
-      BigNumber.from(tokensAmount)
-    );
-
-    return {
-      ...order,
-      hasAllowanceWarning,
-    } as MyOrder;
-  });
-};
